auth-improvements extract error format function to helpers, refactor db connection to async syntax, extract cors check to function
